backend/requirements.txt
fastapi
uvicorn
numpy
reportlab
python-dotenv
sqlalchemy
psycopg2-binary